# Conflicts:
#	public/css/app.css
#	public/css/header.css
#	public/js/app.js
#	public/mix-manifest.json
#	resources/views/public/layouts/footer.blade.php

+++ app/Http/Controllers/WishListController.php
@@ -7,6 +7,8 @@ use Illuminate\Http\Request;
 use App\Http\Requests;
 use App\Http\Controllers\Controller;
 use App\Models\Wishlist;
+use Cartalyst\Sentinel\Native\Facades\Sentinel;
+
 class WishListController extends Controller
 {
 
@@ -18,20 +20,26 @@ class WishListController extends Controller
 
     public function update(Request $request)
     {
-        $user_id = $request->user_id;
-        $product_id = $request->product_id;
-        $action = $request->action;
+        $user = Sentinel::check();
 
-        if($action == 'add'){
-            $to_wishlist = $request->only(['user_id', 'product_id']);
-            Wishlist::create($to_wishlist)->save();
-            $wishlist = Wishlist::where('user_id', $user_id)->count();
-            return response()->json(['count' => $wishlist]);
-        }
-        if($action == 'remove'){
-            Wishlist::where('user_id',$user_id)->where('product_id',$product_id)->delete();
-            $wishlist = Wishlist::where('user_id', $user_id)->count();
-            return response()->json(['count' => $wishlist]);
+        if($user) {
+            $user_id = $user->id;
+            $product_id = (int)$request->product_id;
+            $action = $request->action;
+
+            if($action == 'add'){
+                $to_wishlist = ['user_id' => $user_id, 'product_id' => $product_id];
+                Wishlist::create($to_wishlist)->save();
+                $wishlist = Wishlist::where('user_id', $user_id)->count();
+                return response()->json(['count' => $wishlist]);
+            }
+            if($action == 'remove'){
+                Wishlist::where('user_id',$user_id)->where('product_id',$product_id)->delete();
+                $wishlist = Wishlist::where('user_id', $user_id)->count();
+                return response()->json(['count' => $wishlist]);
+            }
+        }else{
+            return response()->json(['error' => 'login']);
         }
     }
 }
